+++ __tests__/bundle/primitive-types/snapshot.js
@@ -40,9 +40,6 @@ paths:
                     key2: 2
 components: {}
 
-No configurations were defined in extends -- using built in recommended configuration by default.
-Warning! This default behavior is going to be deprecated soon.
-
 Woohoo! Your OpenAPI definitions are valid. 🎉
 
 bundling ./openapi.yaml...