+++ __tests__/lint/assertions-type-integer-in-schema-response/snapshot.js
@@ -1,0 +1,11 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`E2E lint assertions-type-integer-in-schema-response 1`] = `
+
+validating /openapi.yaml...
+/openapi.yaml: validated in <test>ms
+
+Woohoo! Your OpenAPI definition is valid. 🎉
+
+
+`;