+++ packages/core/src/config/__tests__/config-resolvers.test.ts
@@ -1,5 +1,5 @@
 import { colorize } from '../../logger';
-import { asserts } from '../../rules/common/assertions/asserts';
+import { Asserts, asserts } from '../../rules/common/assertions/asserts';
 import { resolveStyleguideConfig, resolveApis, resolveConfig } from '../config-resolvers';
 const path = require('path');
 
@@ -143,7 +143,7 @@ describe('resolveStyleguideConfig', () => {
 
     expect(plugins).toBeDefined();
     expect(plugins?.length).toBe(2);
-    expect(asserts['test-plugin/checkWordsCount']).toBeDefined();
+    expect(asserts['test-plugin/checkWordsCount' as keyof Asserts]).toBeDefined();
   });
 
   it('should throw error when custom assertion load not exist plugin', async () => {
@@ -163,7 +163,7 @@ describe('resolveStyleguideConfig', () => {
       );
     }
 
-    expect(asserts['test-plugin/checkWordsCount']).toBeDefined();
+    expect(asserts['test-plugin/checkWordsCount' as keyof Asserts]).toBeDefined();
   });
 
   it('should correctly merge assertions from nested config', async () => {
@@ -342,7 +342,7 @@ describe('resolveApis', () => {
 });
 
 describe('resolveConfig', () => {
-  it('should add recommended to top level by default', async () => {
+  it('should NOT add recommended to top level by default IF there is a config file', async () => {
     const rawConfig: RawConfig = {
       apis: {
         petstore: {
@@ -373,7 +373,6 @@ describe('resolveConfig', () => {
     expect(apis['petstore'].styleguide.pluginPaths!.map(removeAbsolutePath)).toEqual([]);
 
     expect(apis['petstore'].styleguide.rules).toEqual({
-      ...(await recommendedStyleguidePreset).rules,
       'operation-2xx-response': 'warn',
       'operation-4xx-response': 'error',
     });