+++ packages/core/src/rules/common/assertions/utils.ts
@@ -1,8 +1,15 @@
-import type { AssertResult, RuleSeverity } from '../../../config';
+import { asserts, runOnKeysSet, runOnValuesSet, Asserts } from './asserts';
 import { colorize } from '../../../logger';
 import { isRef, Location } from '../../../ref-utils';
-import { UserContext } from '../../../walk';
-import { asserts } from './asserts';
+import { isTruthy, keysOf, isString } from '../../../utils';
+import type { AssertResult } from '../../../config';
+import type { Assertion, AssertionDefinition, AssertionLocators } from '.';
+import type {
+  Oas2Visitor,
+  Oas3Visitor,
+  SkipFunctionContext,
+  VisitFunction,
+} from '../../../visitors';
 
 export type OrderDirection = 'asc' | 'desc';
 
@@ -11,154 +18,245 @@ export type OrderOptions = {
   property: string;
 };
 
-type Assertion = {
-  property: string | string[];
-  context?: Record<string, any>[];
-  severity?: RuleSeverity;
-  suggest?: any[];
-  message?: string;
-  subject: string;
-};
-
 export type AssertToApply = {
-  name: string;
+  name: keyof Asserts;
   conditions: any;
   runsOnKeys: boolean;
   runsOnValues: boolean;
 };
 
+type AssertionContext = SkipFunctionContext & {
+  node: any;
+};
+
+const assertionMessageTemplates = {
+  problems: '{{problems}}',
+};
+
+function getPredicatesFromLocators(
+  locators: AssertionLocators
+): ((key: string | number) => boolean)[] {
+  const { filterInParentKeys, filterOutParentKeys, matchParentKeys } = locators;
+
+  const keyMatcher = matchParentKeys && regexFromString(matchParentKeys);
+  const matchKeysPredicate =
+    keyMatcher && ((key: string | number) => keyMatcher.test(key.toString()));
+
+  const filterInPredicate =
+    Array.isArray(filterInParentKeys) &&
+    ((key: string | number) => filterInParentKeys.includes(key.toString()));
+
+  const filterOutPredicate =
+    Array.isArray(filterOutParentKeys) &&
+    ((key: string | number) => !filterOutParentKeys.includes(key.toString()));
+
+  return [matchKeysPredicate, filterInPredicate, filterOutPredicate].filter(isTruthy);
+}
+
+export function getAssertsToApply(assertion: AssertionDefinition): AssertToApply[] {
+  const assertsToApply = keysOf(asserts)
+    .filter((assertName) => assertion.assertions[assertName] !== undefined)
+    .map((assertName) => {
+      return {
+        name: assertName,
+        conditions: assertion.assertions[assertName],
+        runsOnKeys: runOnKeysSet.has(assertName),
+        runsOnValues: runOnValuesSet.has(assertName),
+      };
+    });
+
+  const shouldRunOnKeys: AssertToApply | undefined = assertsToApply.find(
+    (assert: AssertToApply) => assert.runsOnKeys && !assert.runsOnValues
+  );
+  const shouldRunOnValues: AssertToApply | undefined = assertsToApply.find(
+    (assert: AssertToApply) => assert.runsOnValues && !assert.runsOnKeys
+  );
+
+  if (shouldRunOnValues && !assertion.subject.property) {
+    throw new Error(
+      `${shouldRunOnValues.name} can't be used on all keys. Please provide a single property`
+    );
+  }
+
+  if (shouldRunOnKeys && assertion.subject.property) {
+    throw new Error(
+      `${shouldRunOnKeys.name} can't be used on a single property. Please use 'property'.`
+    );
+  }
+
+  return assertsToApply;
+}
+
+function getAssertionProperties({ subject }: AssertionDefinition): string[] {
+  return (Array.isArray(subject.property) ? subject.property : [subject?.property]).filter(
+    Boolean
+  ) as string[];
+}
+
+function applyAssertions(
+  assertionDefinition: AssertionDefinition,
+  asserts: AssertToApply[],
+  { rawLocation, rawNode, resolve, location, node }: AssertionContext
+): AssertResult[] {
+  const properties = getAssertionProperties(assertionDefinition);
+  const assertResults: Array<AssertResult[]> = [];
+
+  for (const assert of asserts) {
+    const currentLocation = assert.name === 'ref' ? rawLocation : location;
+
+    if (properties.length) {
+      for (const property of properties) {
+        // we can have resolvable scalar so need to resolve value here.
+        const value = isRef(node[property]) ? resolve(node[property])?.node : node[property];
+        assertResults.push(
+          runAssertion({
+            values: value,
+            rawValues: rawNode[property],
+            assert,
+            location: currentLocation.child(property),
+          })
+        );
+      }
+    } else {
+      const value = assert.name === 'ref' ? rawNode : Object.keys(node);
+      assertResults.push(
+        runAssertion({
+          values: Object.keys(node),
+          rawValues: value,
+          assert,
+          location: currentLocation,
+        })
+      );
+    }
+  }
+
+  return assertResults.flat();
+}
+
 export function buildVisitorObject(
-  subject: string,
-  context: Record<string, any>[],
-  subjectVisitor: any
-) {
-  if (!context) {
-    return { [subject]: subjectVisitor };
+  assertion: Assertion,
+  subjectVisitor: VisitFunction<any>
+): Oas2Visitor | Oas3Visitor {
+  const targetVisitorLocatorPredicates = getPredicatesFromLocators(assertion.subject);
+  const targetVisitorSkipFunction = targetVisitorLocatorPredicates.length
+    ? (node: any, key: string | number) =>
+        !targetVisitorLocatorPredicates.every((predicate) => predicate(key))
+    : undefined;
+  const targetVisitor: Oas2Visitor | Oas3Visitor = {
+    [assertion.subject.type]: {
+      enter: subjectVisitor,
+      ...(targetVisitorSkipFunction && { skip: targetVisitorSkipFunction }),
+    },
+  };
+
+  if (!Array.isArray(assertion.where)) {
+    return targetVisitor;
   }
 
   let currentVisitorLevel: Record<string, any> = {};
   const visitor: Record<string, any> = currentVisitorLevel;
+  const context = assertion.where;
 
   for (let index = 0; index < context.length; index++) {
-    const node = context[index];
-    if (context.length === index + 1 && node.type === subject) {
-      // Visitors don't work properly for the same type nested nodes, so
-      // as a workaround for that we don't create separate visitor for the last element
-      // which is the same as subject;
-      // we will check includes/excludes it in the last visitor.
-      continue;
-    }
-    const matchParentKeys = node.matchParentKeys;
-    const excludeParentKeys = node.excludeParentKeys;
+    const assertionDefinitionNode = context[index];
 
-    if (matchParentKeys && excludeParentKeys) {
+    if (!isString(assertionDefinitionNode.subject?.type)) {
       throw new Error(
-        `Both 'matchParentKeys' and 'excludeParentKeys' can't be under one context item`
+        `${assertion.assertionId} -> where -> [${index}]: 'type' (String) is required`
       );
     }
 
-    if (matchParentKeys || excludeParentKeys) {
-      currentVisitorLevel[node.type] = {
-        skip: (_value: any, key: string) => {
-          if (matchParentKeys) {
-            return !matchParentKeys.includes(key);
-          }
-          if (excludeParentKeys) {
-            return excludeParentKeys.includes(key);
-          }
-        },
+    const locatorPredicates = getPredicatesFromLocators(assertionDefinitionNode.subject);
+    const assertsToApply = getAssertsToApply(assertionDefinitionNode);
+
+    const skipFunction = (
+      node: unknown,
+      key: string | number,
+      { location, rawLocation, resolve, rawNode }: SkipFunctionContext
+    ): boolean =>
+      !locatorPredicates.every((predicate) => predicate(key)) ||
+      !!applyAssertions(assertionDefinitionNode, assertsToApply, {
+        location,
+        node,
+        rawLocation,
+        rawNode,
+        resolve,
+      }).length;
+
+    const nodeVisitor = {
+      ...((locatorPredicates.length || assertsToApply.length) && { skip: skipFunction }),
+    };
+
+    if (
+      assertionDefinitionNode.subject.type === assertion.subject.type &&
+      index === context.length - 1
+    ) {
+      // We have to merge the visitors if the last node inside the `where` is the same as the subject.
+      targetVisitor[assertion.subject.type] = {
+        enter: subjectVisitor,
+        ...((nodeVisitor.skip && { skip: nodeVisitor.skip }) ||
+          (targetVisitorSkipFunction && {
+            skip: (
+              node,
+              key,
+              ctx // We may have locators defined on assertion level and on where level for the same node type
+            ) => !!(nodeVisitor.skip?.(node, key, ctx) || targetVisitorSkipFunction?.(node, key)),
+          })),
       };
     } else {
-      currentVisitorLevel[node.type] = {};
+      currentVisitorLevel = currentVisitorLevel[assertionDefinitionNode.subject?.type] =
+        nodeVisitor;
     }
-    currentVisitorLevel = currentVisitorLevel[node.type];
   }
 
-  currentVisitorLevel[subject] = subjectVisitor;
+  currentVisitorLevel[assertion.subject.type] = targetVisitor[assertion.subject.type];
 
   return visitor;
 }
 
-export function buildSubjectVisitor(
-  assertId: string,
-  assertion: Assertion,
-  asserts: AssertToApply[]
-) {
-  return (
-    node: any,
-    { report, location, rawLocation, key, type, resolve, rawNode }: UserContext
-  ) => {
-    let properties = assertion.property;
-    // We need to check context's last node if it has the same type as subject node;
-    // if yes - that means we didn't create context's last node visitor,
-    // so we need to handle 'matchParentKeys' and 'excludeParentKeys' conditions here;
-    if (assertion.context) {
-      const lastContextNode = assertion.context[assertion.context.length - 1];
-      if (lastContextNode.type === type.name) {
-        const matchParentKeys = lastContextNode.matchParentKeys;
-        const excludeParentKeys = lastContextNode.excludeParentKeys;
-
-        if (matchParentKeys && !matchParentKeys.includes(key)) {
-          return;
-        }
-        if (excludeParentKeys && excludeParentKeys.includes(key)) {
-          return;
-        }
-      }
-    }
-
-    if (properties) {
-      properties = Array.isArray(properties) ? properties : [properties];
-    }
+export function buildSubjectVisitor(assertId: string, assertion: Assertion): VisitFunction<any> {
+  return (node: any, { report, location, rawLocation, resolve, rawNode }) => {
+    const properties = getAssertionProperties(assertion);
 
     const defaultMessage = `${colorize.blue(assertId)} failed because the ${colorize.blue(
-      assertion.subject
-    )}${colorize.blue(
-      properties ? ` ${(properties as string[]).join(', ')}` : ''
-    )} didn't meet the assertions: {{problems}}`;
+      assertion.subject.type
+    )} ${colorize.blue(properties.join(', '))} didn't meet the assertions: ${
+      assertionMessageTemplates.problems
+    }`.replace(/ +/g, ' ');
 
-    const assertResults: Array<AssertResult[]> = [];
-    for (const assert of asserts) {
-      const currentLocation = assert.name === 'ref' ? rawLocation : location;
-      if (properties) {
-        for (const property of properties) {
-          // we can have resolvable scalar so need to resolve value here.
-          const value = isRef(node[property]) ? resolve(node[property])?.node : node[property];
-          assertResults.push(
-            runAssertion({
-              values: value,
-              rawValues: rawNode[property],
-              assert,
-              location: currentLocation.child(property),
-            })
-          );
-        }
-      } else {
-        const value = assert.name === 'ref' ? rawNode : Object.keys(node);
-        assertResults.push(
-          runAssertion({
-            values: Object.keys(node),
-            rawValues: value,
-            assert,
-            location: currentLocation,
-          })
-        );
+    const problems = applyAssertions(assertion, getAssertsToApply(assertion), {
+      rawLocation,
+      rawNode,
+      resolve,
+      location,
+      node,
+    });
+
+    if (problems.length) {
+      for (const problemGroup of groupProblemsByPointer(problems)) {
+        const message = assertion.message || defaultMessage;
+        const problemMessage = getProblemsMessage(problemGroup);
+        report({
+          message: message.replace(assertionMessageTemplates.problems, problemMessage),
+          location: getProblemsLocation(problemGroup) || location,
+          forceSeverity: assertion.severity || 'error',
+          suggest: assertion.suggest || [],
+          ruleId: assertId,
+        });
       }
     }
-
-    const problems = assertResults.flat();
-    if (problems.length) {
-      const message = assertion.message || defaultMessage;
-
-      report({
-        message: message.replace('{{problems}}', getProblemsMessage(problems)),
-        location: getProblemsLocation(problems) || location,
-        forceSeverity: assertion.severity || 'error',
-        suggest: assertion.suggest || [],
-        ruleId: assertId,
-      });
-    }
   };
+}
+
+function groupProblemsByPointer(problems: AssertResult[]): AssertResult[][] {
+  const groups: Record<string, AssertResult[]> = {};
+  for (const problem of problems) {
+    if (!problem.location) continue;
+    const pointer = problem.location.pointer;
+    groups[pointer] = groups[pointer] || [];
+    groups[pointer].push(problem);
+  }
+  return Object.values(groups);
 }
 
 function getProblemsLocation(problems: AssertResult[]) {